Change Current library to interpret LowPrecisionCurrent and HighPrecisionCurrent as mA. Prior to this, some functions interpreted both in mA and others interpreted them in A, causing inconsistent measurements and readings to appear

+++ BeVolt/Apps/Inc/Current.h
@@ -37,14 +37,14 @@ int8_t Current_IsCharging(void);
 
 /** Current_GetHighPrecReading
  * Gets the Ampere measurement the high precision hall effect sensor recorded
- * @return Amperes value
+ * @return milliamperes value
  */
-int16_t Current_GetHighPrecReading(void);
+int32_t Current_GetHighPrecReading(void);
 
 /** Current_GetLowPrecReading
  * Gets the Ampere measurement the low precision hall effect sensor recorded
- * @return Amperes value
+ * @return milliamperes value
  */
-int16_t Current_GetLowPrecReading(void);
+int32_t Current_GetLowPrecReading(void);
 
 #endif
